CB-13145: pass full options to plugman uninstall
