+++ src/cordova/plugin/add.js
@@ -99,28 +99,8 @@ function add (projectRoot, hooksRunner, opts) {
                 }).then(function (directory) {
                     return pluginInfoProvider.get(directory);
                 }).then(function (pluginInfo) {
-                    // Validate top-level required variables
-                    var pluginVariables = pluginInfo.getPreferences();
-                    opts.cli_variables = opts.cli_variables || {};
-                    var pluginEntry = cfg.getPlugin(pluginInfo.id);
-                    // Get variables from config.xml
-                    var configVariables = pluginEntry ? pluginEntry.variables : {};
-                    // Add config variable if it's missing in cli_variables
-                    Object.keys(configVariables).forEach(function (variable) {
-                        opts.cli_variables[variable] = opts.cli_variables[variable] || configVariables[variable];
-                    });
-                    var missingVariables = Object.keys(pluginVariables)
-                        .filter(function (variableName) {
-                            // discard variables with default value
-                            return !(pluginVariables[variableName] || opts.cli_variables[variableName]);
-                        });
-
-                    if (missingVariables.length) {
-                        events.emit('verbose', 'Removing ' + pluginInfo.dir + ' because mandatory plugin variables were missing.');
-                        shell.rm('-rf', pluginInfo.dir);
-                        var msg = 'Variable(s) missing (use: --variable ' + missingVariables.join('=value --variable ') + '=value).';
-                        return Q.reject(new CordovaError(msg));
-                    }
+                    
+                    plugin_util.mergeVariables(pluginInfo, cfg, opts);
 
                     // Iterate (in serial!) over all platforms in the project and install the plugin.
                     return chainMap(platformList, function (platform) {
